contemplant refactor
