+++ hierophant/contemplant.example.toml
@@ -1,1 +1,16 @@
-#TODO:
+# REQUIRED
+
+# websocket address in the form of <ws://127.0.0.1:3000/ws>
+# The port is usually the same as the Hierophant http port.
+hierophant_ws_address = "ws://127.0.0.1:3000/ws"
+
+# OPTIONAL
+# Below is a list of all config variables
+# See hierophant/contemplant/config.rs for default values and explanations.
+# contemplant_name =
+# moongate_endpoint =
+# heartbeat_interval_seconds =
+# max_proofs_stored =
+# magister_drop_endpoint =
+# assessor.moongate_log_path =
+# assessor.watcher_polling_interval_ms =
