+++ hierophant/contemplant/src/proof_executor/mod.rs
@@ -1,0 +1,4 @@
+mod assessor;
+mod executor;
+
+pub use executor::execute_proof;
