+++ hierophant/contemplant/src/proof_store/mod.rs
@@ -1,0 +1,5 @@
+mod client;
+mod command;
+mod store;
+
+pub use client::ProofStoreClient;
